requirements.txt
django~=4.0
djangorestframework
pyyaml
requests
django-cors-headers
pymongo
Pillow
drf_yasg
pep8
djangorestframework-simplejwt
pyjwt